Refactor: 파일명 및 URI명 변경
