+++ app.js
@@ -7,14 +7,14 @@ dotenv.config();
 
 app.listen(process.env.PORT);
 
-const userRouter = require('./routes/users');
+const authRouter = require('./routes/auth');
 const bookRouter = require('./routes/books');
 const categoryRouter = require('./routes/category');
 const likeRouter = require('./routes/likes');
 const cartRouter = require('./routes/carts');
 const orderRouter = require('./routes/orders');
 
-app.use('/users', userRouter);
+app.use('/auth', authRouter);
 app.use('/books', bookRouter);
 app.use('/category', categoryRouter);
 app.use('/likes', likeRouter);
