+++ routes/auth.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
-const { signupRequest, signin, pwdResetRequest, pwdReset, signupConfirm } = require('../controller/UserController');
+const { signupRequest, signin, pwdResetRequest, pwdReset, signupConfirm } = require('../controller/AuthController');
 
-const { validatesSignup, validatesSignin, validatesEmail } = require('../middleware/userValidator');
+const { validatesSignup, validatesSignin, validatesEmail } = require('../middleware/authValidator');
 
 router.use(express.json());
